tsparser: correctly parse pubsub attrs
