+++ tsparser/src/parser/types/type_string.rs
@@ -163,6 +163,7 @@ where
         match &s.location {
             super::WireLocation::Query => self.buf.write_str("Query<")?,
             super::WireLocation::Header => self.buf.write_str("Header<")?,
+            super::WireLocation::PubSubAttr => self.buf.write_str("Attribute<")?,
         }
         self.render_type(&s.underlying)?;
         if let Some(name) = &s.name_override {
